+++ src/Hawkeye.Core/ComponentModel/ComponentModelExtensions.cs
@@ -40,15 +40,16 @@ namespace Hawkeye.ComponentModel
         public static PropertyDescriptorCollection GetAllProperties(
             this ITypeDescriptorContext context,
             object component,
-            Attribute[] attributes,
-            bool inspectBaseClasses = true)
+            bool inspectBaseClasses = true,
+            bool retrieveStaticMembers = true,
+            bool keepOriginalCategoryAttribute = true)
         {
             if (component == null || component.GetType().IsPrimitive || component is string)
                 return new PropertyDescriptorCollection(new PropertyDescriptor[] { });
-            
+
             // Make sure we are inspecting the real component
             component = component.GetInnerObject();
-            
+
             var type = component.GetType();
 
             var allprops = new Dictionary<string, PropertyDescriptor>();
@@ -62,8 +63,8 @@ namespace Hawkeye.ComponentModel
                     var fullName = pi.DeclaringType.FullName + "." + pi.Name;
                     if (excludedProperties.Contains(fullName)) return;
 
-                    if (isStatic) allprops.Add(pi.Name, new StaticPropertyDescriptor(type, pi));
-                    else allprops.Add(pi.Name, new InstancePropertyDescriptor(component, type, pi));
+                    if (isStatic) allprops.Add(pi.Name, new StaticPropertyPropertyDescriptor(type, pi, keepOriginalCategoryAttribute));
+                    else allprops.Add(pi.Name, new InstancePropertyPropertyDescriptor(component, type, pi, keepOriginalCategoryAttribute));
                 }
                 catch (Exception ex)
                 {
@@ -78,8 +79,9 @@ namespace Hawkeye.ComponentModel
                 foreach (var pi in type.GetProperties(instanceFlags))
                     addPropertyDescriptor(pi, false);
 
-                foreach (var pi in type.GetProperties(staticFlags))
-                    addPropertyDescriptor(pi, true);
+                if (retrieveStaticMembers)
+                    foreach (var pi in type.GetProperties(staticFlags))
+                        addPropertyDescriptor(pi, true);
 
                 if (type == typeof(object) || !inspectBaseClasses)
                     break;
@@ -87,56 +89,16 @@ namespace Hawkeye.ComponentModel
                 depth++;
             }
             while (true);
-            
+
             return new PropertyDescriptorCollection(allprops.Values.ToArray());
-
-            //PropertyDescriptorCollection properties;
-            //PropertyDescriptorCollection childElements = null;
-
-            ////Attribute[] attributeArray1 = new Attribute[] { };
-            ////attributes = attributeArray1; // replace the attribute array to allow all properties to be visible.
-
-            //if (component is IEnumerable)
-            //    childElements = ConverterUtils.GetEnumerableChildsAsProperties(context, component, attributes);
-
-            //if (context == null)
-            //{
-            //    //properties = TypeDescriptor.GetProperties(component, attributes);
-            //    properties = GetAllPropertiesCustom(component, attributes);
-            //}
-            //else
-            //{
-            //    RemapPropertyDescriptor remapDescriptor = context.PropertyDescriptor as RemapPropertyDescriptor;
-            //    if (remapDescriptor != null)
-            //    {
-            //        component = remapDescriptor.OriginalComponent;
-            //    }
-
-            //    TypeConverter converter1 = (context.PropertyDescriptor == null) ? 
-            //        TypeDescriptor.GetConverter(component) : context.PropertyDescriptor.Converter;
-            //    if ((converter1 != null) && converter1.GetPropertiesSupported(context))
-            //    {
-            //        //					properties = converter1.GetProperties(context, component, attributes);
-            //        //					if ( properties == null )
-            //        properties = GetAllPropertiesCustom(component, attributes);
-            //    }
-            //    else
-            //    {
-            //        properties = GetAllPropertiesCustom(component, attributes);
-            //    }
-            //}
-
-            //if (childElements != null && childElements.Count > 0)
-            //    properties = MergeProperties(properties, childElements);
-
-            //return properties;
         }
 
         public static PropertyDescriptorCollection GetAllEvents(
             this ITypeDescriptorContext context,
             object component,
-            Attribute[] attributes,
-            bool inspectBaseClasses = true)
+            bool inspectBaseClasses = true,
+            bool retrieveStaticMembers = true,
+            bool keepOriginalCategoryAttribute = true)
         {
             if (component == null || component.GetType().IsPrimitive || component is string)
                 return new PropertyDescriptorCollection(new PropertyDescriptor[] { });
@@ -157,8 +119,8 @@ namespace Hawkeye.ComponentModel
                     var fullName = ei.DeclaringType.FullName + "." + ei.Name;
                     if (excludedEvents.Contains(fullName)) return;
 
-                    if (isStatic) allevs.Add(ei.Name, new StaticEventPropertyDescriptor(type, ei));
-                    else allevs.Add(ei.Name, new InstanceEventPropertyDescriptor(component, type, ei));
+                    if (isStatic) allevs.Add(ei.Name, new StaticEventPropertyDescriptor(type, ei, keepOriginalCategoryAttribute));
+                    else allevs.Add(ei.Name, new InstanceEventPropertyDescriptor(component, type, ei, keepOriginalCategoryAttribute));
 
                 }
                 catch (Exception ex)
@@ -174,8 +136,9 @@ namespace Hawkeye.ComponentModel
                 foreach (var ei in type.GetEvents(instanceFlags))
                     addPropertyDescriptor(ei, false);
 
-                foreach (var ei in type.GetEvents(staticFlags))
-                    addPropertyDescriptor(ei, true);
+                if (retrieveStaticMembers)
+                    foreach (var ei in type.GetEvents(staticFlags))
+                        addPropertyDescriptor(ei, true);
 
                 if (type == typeof(object) || !inspectBaseClasses)
                     break;
@@ -187,6 +150,14 @@ namespace Hawkeye.ComponentModel
 
             return new PropertyDescriptorCollection(allevs.Values.ToArray());
         }
+
+        ////public static PropertyDescriptorCollection GetInstanceEvents(
+        ////    this ITypeDescriptorContext context,
+        ////    object component, 
+        ////    bool keepOriginalCategoryAttribute = true)
+        ////{
+        ////    return GetAllEvents(context, component, true, false, keepOriginalCategoryAttribute);
+        ////}
 
         #endregion